third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-customization/scrollstate-consume-deltas-throw.html

<!DOCTYPE html>
<html>
<head>
<title>ScrollState's consumeDelta method throws on invalid deltas</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

test(function() {
  assert_true('ScrollState' in window, "'ScrollState' in window");
}, "These tests only work with scroll customization enabled.");

function testConsumingDelta(dx, dy, consume_x, consume_y, testName, expectedError) {
  test(function() {
    var scrollState = new ScrollState({deltaX: dx, deltaY: dy});
    assert_throws('InvalidModificationError', function() {
      scrollState.consumeDelta(consume_x, consume_y);
    });
    assert_equals(scrollState.deltaX, dx);
    assert_equals(scrollState.deltaY, dy);
  }, testName);
}

if ('ScrollState' in window) {
  // Consume from an event with a delta of (5, 5).
  testConsumingDelta(5, 5, 6, 4, "Overconsumed X");
  testConsumingDelta(5, 5, 4, 6, "Overconsumed Y");
  testConsumingDelta(5, 5, -1, 4, "Can't increase X delta");
  testConsumingDelta(5, 5, 4, -1, "Can't increase Y delta");

  // Consume from an event with a delta of (-5, -5).
  testConsumingDelta(-5, -5, -6, -4, "Overconsumed X with negative delta");
  testConsumingDelta(-5, -5, -4, -6, "Overconsumed Y with negative delta");
  testConsumingDelta(-5, -5, 1, -4, "Can't increase X delta magnitude");
  testConsumingDelta(-5, -5, -4, 1, "Can't increase Y delta magnitude");
}
</script>
</body>
</html>
